Add getFilteredParameterMappingsForDashcardText helper
