+++ frontend/src/metabase/dashboard/actions/parameters.tsx
@@ -3,7 +3,6 @@ import { assoc } from "icepick";
 import { t } from "ttag";
 import _ from "underscore";
 
-import { tag_names } from "cljs/metabase.parameters.shared";
 import CS from "metabase/css/core/index.css";
 import { showAutoWireToast } from "metabase/dashboard/actions/auto-wire-parameters/actions";
 import {
@@ -17,6 +16,7 @@ import { createAction, createThunkAction } from "metabase/lib/redux";
 import {
   type NewParameterOpts,
   createParameter,
+  getFilteredParameterMappingsForDashcardText,
   setParameterName as setParamName,
   setParameterType as setParamType,
 } from "metabase/parameters/utils/dashboards";
@@ -28,7 +28,6 @@ import {
   PULSE_PARAM_EMPTY,
   isParameterValueEmpty,
 } from "metabase-lib/v1/parameters/utils/parameter-values";
-import { getTextTagFromTarget } from "metabase-lib/v1/parameters/utils/targets";
 import type {
   ActionDashboardCard,
   CardId,
@@ -459,29 +458,22 @@ export const updateParameterMappingsForDashcardText = createThunkAction(
       const parameterMappings = dashcard?.parameter_mappings;
       const text = dashcard?.visualization_settings?.text;
 
-      if (!parameterMappings?.length || text === null || text === undefined) {
+      if (typeof text !== "string") {
         return;
       }
 
-      const tagNames = tag_names(text);
-      const filteredParameterMapping = parameterMappings.filter((mapping) => {
-        const target = mapping.target;
+      const filteredParameterMappings =
+        getFilteredParameterMappingsForDashcardText(parameterMappings, text);
 
-        const textTag = getTextTagFromTarget(target);
-
-        // A different tag type (not a text-tag) or no tag at all means we keep the mapping
-        if (!textTag) {
-          return true;
-        }
-
-        return tagNames.includes(textTag);
-      });
+      if (!filteredParameterMappings) {
+        return;
+      }
 
       dispatch(
         setDashCardAttributes({
           id: dashcardId,
           attributes: {
-            parameter_mappings: filteredParameterMapping,
+            parameter_mappings: filteredParameterMappings,
           },
         }),
       );
