+++ frontend/src/metabase/parameters/utils/dashboards.ts
@@ -1,5 +1,6 @@
 import _ from "underscore";
 
+import { tag_names } from "cljs/metabase.parameters.shared";
 import { isQuestionCard, isQuestionDashCard } from "metabase/dashboard/utils";
 import { slugify } from "metabase/lib/formatting";
 import { isNotNull } from "metabase/lib/types";
@@ -13,6 +14,7 @@ import type {
 import { isFieldFilterParameter } from "metabase-lib/v1/parameters/utils/parameter-type";
 import {
   getParameterTargetField,
+  getTextTagFromTarget,
   isParameterVariableTarget,
 } from "metabase-lib/v1/parameters/utils/targets";
 import type {
@@ -324,3 +326,27 @@ export function getFilteringParameterValuesMap(
 
   return filteringParameterValues;
 }
+
+export const getFilteredParameterMappingsForDashcardText = (
+  parameterMappings: DashboardCard["parameter_mappings"],
+  dashcardText: string,
+) => {
+  if (!parameterMappings) {
+    return parameterMappings;
+  }
+
+  const tagNames = tag_names(dashcardText);
+
+  return parameterMappings.filter((mapping) => {
+    const target = mapping.target;
+
+    const textTag = getTextTagFromTarget(target);
+
+    // A different tag type (not a text-tag) or no tag at all means we keep the mapping
+    if (!textTag) {
+      return true;
+    }
+
+    return tagNames.includes(textTag);
+  });
+};
